pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sdd"
version = "0.8.0"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=0.10",
    "PyYAML>=6.0",
    "click>=8.0",
]

[project.scripts]
sdd = "sdd.cli:main"
sdd-hook-log = "sdd.hooks.log_tool:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
explicit_package_bases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--tb=short"

[tool.coverage.run]
source = ["sdd"]
branch = true

[tool.coverage.report]
fail_under = 80
